feat(conversation_rendering_v1): move current implementation (no-op)
